+++ inc/minishell.h
@@ -6,7 +6,7 @@
 /*   By: tpauvret <tpauvret@student.42lausanne.ch>  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/06 21:52:16 by tpauvret          #+#    #+#             */
-/*   Updated: 2022/03/07 22:29:01 by tpauvret         ###   ########.fr       */
+/*   Updated: 2022/03/07 22:58:16 by tpauvret         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -186,15 +186,9 @@ int				check_wrong_char(char *str);
 char			*join_path_to_arg(char *path, char *arg);
 int				exec_program(t_command *child, t_env **env_list);
 int				is_builtins(t_env **env_list, t_command *child);
-int				execution(t_env **env_list, t_mini *shell);
+int				process_cmd(t_env **env_list, t_mini *shell);
 
-// EXECUTION - redirection.c
-int				redirection(t_redir *redirection, t_process *proc);
-int				db_great_than(int *fd, t_redir *redirection);
-int				great_than(int *fd, t_redir *redirection);
-int				less_than(int *fd, t_redir *redirection);
-
-// EXCUTION - exec_tools.c
+// EXECUTION - exec_tools.c
 int				print_error(char *str, char *msg, int ret);
 char			**args_to_array(t_command *child);
 char			**lst_to_array(t_env **env_list);
@@ -209,13 +203,6 @@ int				check_path(t_env **env_list, t_command *child);
 int				find_file(t_env **env_list, t_command *child);
 int				file_error(t_redir *redirection);
 
-// EXECUTION - pipe.c
-int				pipe_my_ride(t_mini *shell, t_process *proc, t_env **env_list);
-int				close_fd(t_mini *shell, t_process *proc);
-int				create_fd_pipe(t_mini *shell, t_process *proc);
-int				pipe_fork(t_command	*child, t_env **env_list,
-					t_process *proc, int *cmd);
-
 //EXECUTION - heredoc.c
 void			exec_db_less(char *stop, int *heredoc_fd);
 int				redir_heredocs(t_redir *redir, int fd, bool last);
@@ -226,6 +213,27 @@ void			process_heredocs(t_mini *shell);
 //EXECUTION - heredocs_utils.c
 void			newline(int signal);
 void			stop_heredoc(int signal);
+
+// EXECUTION - redirection.c
+bool			redir_output(t_redir *redir);
+bool			redir_append(t_redir *redir);
+bool			redir_input(t_redir *redir);
+bool			redirect(t_redir *redir, int *pipe_fd);
+bool			exec_redirections(t_redir *redir, int *pipe_fd);
+
+// EXECUTION - pipe.c
+void			init_pipes(t_command *tmp);
+void			left_pipe(t_command *cmds);
+void			right_pipe(t_command *cmds);
+bool			ft_lstall(t_redir *lst, bool (*f)(void *));
+bool			ft_lstany(t_redir *lst, bool (*f)(t_redir *));
+
+// EXECUTION - op_control.c
+bool			is_stdout_redir(t_redir *redir);
+void			fd_reset(t_mini *shell);
+void			close_pipes(t_command *child);
+bool			input_file_exist(void *redir_ptr);
+bool			op_control(t_command *child);
 
 // Signal - signal.c
 void			run_signals(int sig);
@@ -331,9 +339,5 @@ void			unset(t_built_args *args, t_env **env_list);
 void			check_digit(char *arg);
 int				check_exit(char *arg);
 int				exit_func(t_built_args *args);
-
-
-//TEST
-int				process_cmd(t_env **env_list, t_mini *shell);
 
 #endif
